requirements.txt
asgiref
black
certifi
cffi
charset-normalizer
click
colorama
cryptography
dj-database-url
Django
django-cors-headers
django-environ
django-paypal
gunicorn=
idna
mypy-extensions
packaging
pathspec
paypal-checkout-serversdk
paypalhttp
paypalrestsdk
pillow
platformdirs
psycopg2
psycopg2-binary
pycparser
pyOpenSSL
python-dotenv
pytz
requests
ruff
six
sqlparse
typing_extensions
tzdata
urllib3
whitenoise
dj-database-url

